requirements.txt
# Die wichtigste Flask-Erweiterungen für die Webanwendung
flask~=3.0.0
flask-bcrypt~=1.0.1
flask-sqlalchemy~=3.1.1
psycopg2-binary
requests~=2.31.0
SQLAlchemy~=2.0.25
gunicorn
